feat: implement canonical AST architecture foundation
- Create unified CanonicalAispDocument type system in ast/canonical.rs
- Implement IntoCanonical trait for lossless type conversion
- Add robust parser integration with canonical conversion
- Establish single source of truth for AISP document types
- Eliminate dual AST architecture (legacy vs robust parser)

Breaking Changes:
- Unified all AISP types under canonical:: namespace
- Deprecated direct use of robust_parser:: types in favor of canonical conversion

Impact: Resolves type system fragmentation affecting 48 modules
Addresses: ADR-027 Canonical AST Architecture

+++ aisp-formal-verification/crates/aisp-core/src/parser/robust_parser.rs
@@ -697,7 +697,7 @@ impl RobustAispParser {
 
     /// Find balanced delimiter pairs
     fn find_balanced_delimiter(&self, text: &str, open: char, close: char) -> Option<usize> {
-        let mut depth = 0;
+        let mut depth = 0i32;
         let mut start_found = false;
 
         for (i, ch) in text.char_indices() {
@@ -800,14 +800,14 @@ impl RobustAispParser {
     }
 
     fn calculate_nesting_depth(&self, input: &str) -> usize {
-        let mut depth = 0;
-        let mut max_depth = 0;
+        let mut depth = 0i32;
+        let mut max_depth = 0usize;
 
         for ch in input.chars() {
             match ch {
                 '{' | '⟨' | '(' => {
                     depth += 1;
-                    max_depth = max_depth.max(depth);
+                    max_depth = max_depth.max(depth as usize);
                 }
                 '}' | '⟩' | ')' => {
                     depth = depth.saturating_sub(1);
